When handling a BUFFER_UNDERFLOW after a unwrap, do not close engine inbound.
Closing the engine after reading a bunch of bytes will cause problems when trying to use the engine to unwrap those bytes, that are still in the buffer, wrapped.
Also, it could also be the case that the engine has to be used to wrap a close message. The SSLEngineImpl underneath SSLEngine has only one open state, and closing
inbound will cause outbound to be closed as well.